requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# API and HTTP
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Secrets Management
hvac==1.2.1
boto3==1.34.0  # For AWS Secrets Manager
azure-keyvault-secrets==4.7.0

# Graph Database
neo4j==5.15.0
gremlinpython==3.7.1  # For Neptune

# Message Queue
kafka-python==2.0.2
nats-py==2.6.0

# OCSF and Data Processing
jsonschema==4.20.0
pyyaml==6.0.1

# Agentic Core
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2

# Scheduling
apscheduler==3.10.4
celery==5.3.4

# Web Framework (for onboarding wizard)
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Monitoring and Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.25.2  # For testing

# Utilities
python-dateutil==2.8.2
pytz==2024.1
structlog==24.1.0

# Circuit Breaker
pybreaker==1.0.1

# Plugin System
pluggy==1.4.0

